test: fold the watch-options tests into file-watcher.test.ts
A separate test file with no matching source file was a smell. Merge the
chokidar-options assertions back into file-watcher.test.ts using
vi.mock("chokidar", { spy: true }): spy mode keeps the real implementation so
the existing integration tests still watch real temp dirs, while the new
"watch options" suite overrides watch() per-test (vi.mocked + mockImplementation)
to inspect the options object, then mockRestore hands the real, spied watch
back. Gating is still mutation-verified through the spied tests.
